requirements.txt
Pillow
tk
pytweening
numpy
paho-mqtt
loguru
pyperclip
beautifulsoup4
requests